Phase 2.1: move CaffNodeConfig and IsEspressoEnforcement to sibling file
Reduces op-node/rollup/types.go espresso-specific diff from 60 to 30 LOC.

The CaffNodeConfig struct definition (27 lines, including doc comment)
and the IsEspressoEnforcement method (8 lines) move out of types.go
into a new sibling file op-node/rollup/espresso_types.go. Pure code
move with zero behavior change.

The new file has no build tag because Config is used in both mips64
(op-program-client-mips64) and !mips64 builds. The existing
espresso_config.go (which holds CaffNodeConfig.ToCLIConfig and
CaffNodeConfigFromCLIConfig) remains !mips64-tagged because it imports
the espresso package which is not mips64-buildable.

The four espresso field declarations (EspressoEnforcementTime,
CaffNodeConfig, BatchAuthenticatorAddress, BatchAuthLookbackWindow) and
the forEachFork espresso hook stay in types.go: Go has no mechanism to
declare struct fields outside the struct's source file, and grouping
them into an Espresso *EspressoConfig sub-struct would change the
rollup.json on-disk layout (rejected to preserve compatibility with
deployed rollup.json files).
